Refactor claim position handling to use float32 coordinates
Updated claim location struct to use Pos1 and Pos2 fields and switched claim lookup logic to use float32 coordinates for improved precision. Adjusted all usages of ClaimAt to match the new signature and improved bounding box calculation for claims.

+++ gobds/session/handlers/player_auth_input.go
@@ -115,7 +115,7 @@ func (h *PlayerAuthInput) handleClaims(c interceptor.Client, pkt *packet.PlayerA
 			return
 		}
 		pos := action.BlockPos
-		cl := ClaimAt(dat.Dimension(), pos.X(), pos.Z())
+		cl := ClaimAt(dat.Dimension(), float32(pos.X()), float32(pos.Z()))
 
 		if cl.ID == "" || // Invalid claim?
 			cl.OwnerXUID == "*" || // Admin claim.
